requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
google-generativeai==0.3.1
PyPDF2==3.0.1
pandas==2.1.1
python-dotenv==1.0.0
pydantic==2.4.2